+++ test/remote_sources/github_release_test.rb
@@ -3,11 +3,19 @@ require "test_helper"
 module Torba
   class GithubReleaseTest < Minitest::Test
     def remote
-      RemoteSources::GithubRelease.new("repo/user", "v.2.0.0")
+      RemoteSources::GithubRelease.new("user/repo", "v.2.0.0")
     end
 
     def test_source
-      assert_equal "repo/user", remote.source
+      assert_equal "user/repo", remote.source
+    end
+
+    def test_repository_name
+      assert_equal "repo", remote.repository_name
+    end
+
+    def test_repository_user
+      assert_equal "user", remote.repository_user
     end
 
     def test_tag
@@ -15,7 +23,7 @@ module Torba
     end
 
     def test_url
-      assert_equal "https://github.com/repo/user/archive/v.2.0.0.zip", remote.url
+      assert_equal "https://github.com/user/repo/archive/v.2.0.0.zip", remote.url
     end
   end
 end